move all courses routes to separate file and refactore code with Router and module.exports
